Run examples in test
This commit adds a test case for running each of the examples as both a
single deploy and multi-deploy. This test should ensure that the
examples are always in a working state, and provide some end-to-end
coverage of the single and multi deployers.

It also provides some test coverage of the integration of codegen with a
real application.