Fixed `weaver multi` and `weaver ssh` main bug.
Before, `weaver multi` and `weaver ssh` assumed the name of the main
component was "main". Recently, we switched it to
"github.com/ServiceWeaver/weaver/main". Both the meaver multi and weaver
ssh deployers would instruct a weavelet to start "main", but no such
component was registered, and the app would immediately crash.