Embed deployer API version in binary.
Recall that a weavelet communicates with a deployer over a pair of pipes
using the deployer API. Every time the deployer API changes, we assign
it a new version. When a deployer launches a Service Weaver binary, the
deployer checks that the weavelet is running a compatible version.

This PR embeds the deployer API version in Service Weaver binaries,
allowing a deployer to check the version of a binary without running it.
In some cases, this is very convenient. For example, the "weaver gke"
tool can check the version of a binary before submitting it to GKE. The
GKE controller can also check the version without having to run
potentially dangerous user code.

The mechanism to embed and extract versions is the same as the one in
PR #316. A string of the form ⟦107c095e:wEaVeRvErSiOn:0.8.0⟧ is included
in the generated code. The version.ReadVersion function parses this
string from the RO data section of a binary.