Compile time codegen API version checks.
This PR introduces code to check, at compile time, that every
weaver_gen.go file uses the codegen API version linked into the
developer's binary. See runtime/codegen/version.go for more context and
an explanation of how the generated code works.

Note that there are many simpler ways to produce a compiler error when
versions mismatch. We can embed version info into the name of a type
[1]. We can use a duplicate key trick [2] I learned from [4]. We can use
the array trick in this PR without anything more fancy [3]. However,
these approaches don't produce helpful error messages. Something like
the following is a bit cryptic:

    cannot use [0][9][0]int{} (value of type [0][9][0]int) as [0][10][0]int value in variable declaration

The approach in this PR is nice because it includes the generated error
string in the compile error. This way, users can figure out what went
wrong, or at least copy and paste the error message to us so we can
figure out what went wrong. This better error message does introduce
complexity, but thankfully the complexity is hidden away in the codegen
package which users do not interact with directly.

We can alternatively check versions at runtime. An implementation of
that is in PR #327. This approach is simple and produces nice error
messages, but the errors happen at runtime instead of compile time. For
some deployers, like the GKE deployer, waiting to fail at runtime can
be very annoying.

[1]: https://go.dev/play/p/w51PCV9LTPr
[2]: https://go.dev/play/p/vWMbtvboe0s
[3]: https://go.dev/play/p/ebfXpODWOy2
[4]: https://groups.google.com/g/golang-nuts/c/w1-JQMaH7c4