Make weavertest handle benchmarks.
weavertest.Run expects either *testing.T or *testing.B which indicates
that it should work for benchmarking components. However, the
weavertest implementation was always passing a -test.run argument. We
now set the -test.bench argument instead (and disable all tests) if we
detect that the argument passed to weavetest.Run is a *testing.B.