Removed logging package's dependency on testing.
As suggested by Sanjay in #330, this PR removes the logging package's
dependency on the testing package. It does so by replacing testing.TB
with a smaller logging.TB interface that includes the Log and Cleanup
methods. It also receives the result of testing.Verbose as an argument,
rather than calling testing.Verbose directly.