Added WithConfig-related "weaver generate" checks.
Recall that a component implementation struct can embed
`weaver.WithConfig[T]` to access a config of type `T` parsed from a
weaver.toml file. Before this PR, `weaver generate` would generate a
`ConfigFn` function for component structs with an embedded `WithConfig`.
Here's an example:

```
ConfigFn: func(i any) any { return i.(*sqlStore).WithConfig.Config() }
```

When a weavelet constructed a component, it would use this function
(registered through the registry) to populate the config.

However, imagine you embed a `weaver.WithConfig[T]` inside a struct but
then forget to run `weaver generate`. The code will build and run but
because there is no generated `ConfigFn`, the config will not be parsed.

This PR removes the `ConfigFn` as it is not needed. Instead a weavelet
can use reflection directly to decide whether a component struct has an
embedded `weaver.WithConfig[T]`.