Request fresh certificates on every connection establishment. (#347)
This is required as certificates in general have a short validity
and typically get rotated. Previously, we would pass the certificate
to the weavelet and use it indefinitely. We now request a new
(i.e., latest) certificate from the weavelet each time a connection
is established, ensuring that the freshest certificate is always used.