Move app initialization code to a weaver.Main method. (#359)
* Move app initialization code to a weaver.Main method.

Previously, application startup code was wrapped up in a function
passed to weaver.Run. However this made it hard to test startup
code. We now have a `Main` method in `weaver.Main` and place the
startup code in the implementation of this method.

Refactored startup the internal API used by weavertest to separate out
starting the application from running its startup code.

* Disallow "weaver.Ref[weaver.Main]" component fields.

* Drop TODO and add comment for Main