Fixed bug in logic deciding when to run main.
Previously, a weavelet decided to run the main component's Main method
if it had a registered Main component that was marked local. This worked
for the single process deployers but broke all other deployers.

For weavelets not started in singleprocess mode, the weavelet receives
an explicit RunMain field in EnvelopeOptions that signals it to start
running main.

This bug also re-emphasizes the importance of tests that automatically
run all the deployers. I'll prioritize this and finish up PR #287.