Added test to catch incorrect version updates.
Recall that we hardcode the deployer API version into every weaver
application binary. The hardcoded string is called version.
Additionally, we have constants for the Major, Minor, and Patch of the
deployer API version. These two need to stay in sync, but because
version has to be hardcoded, there is no way to programmatically keep
them in sync. This makes it easy to change Major, Minor, and Patch but
forget to update version. This PR adds a unit test that fails if you
forget to change version.