Propagate the internal listener port
In the current implementation, every weavelet will listen for internal
traffic at an arbitrary port. However, for some deployers (e.g.,
kubernetes based deployer), we will prestart the listeners hence the
port has to be propagated from the deployer.

This PR adds a new field `internal_port` that contains the port number
for the internal listener. By default this is 0, which preserves the
current behavior. However, in a future PR, the `internal_port` value
will be used by the kubernetes deployer.