Only push website changes when a tag is created.
Before this PR, every commit merged into main would trigger an update to
our public website. This meant that the docs would reference code that
wasn't yet merged into a stable release. This led to a lot of confusion
as people tried to follow along with the docs, but nothing worked.

This PR changes the GitHub Action to only update the website when we
make a new tag for a release. You can also manually trigger the action
if you want. In the future, we should consider versioning our
documentation and allowing people to view old versions?