Add instrumentation scope to the read span. (#378)
The ReadOnlySpan has a new field `InstrumentationScope` that is supposed
to replace the `InstrumentationLibrary` that is deprecated.

However, our internal implementation of the span doesn't override this
field, hence the exported spans have it nil. When using the jaeger exporter,
it relies on the instrumentation scope to get some info about the spans.

This PR overrides the instrumentation scope field as well, in our
internal span implementation.