Modify the listener format for local listeners. (#394)
This fixes the weirdness of the "local_address" field, which
is only used by multi/single/ssh deployers, but not others
(e.g., GKE, kube). The plan is to move the listener options
into the deployer config sections, since they are largely
different.

Based on discussions with @rgrandl and @mwhittaker.

Other changes:
* Simplify the parsing of the listener config, by reading
  into a `map[string]ListenerOptions`. This obviates the
  need for using the `toml.Primitive` type.

TODO: Change the format of the GKE/kube deployers to be similar.