Fixed buggy `weaver version` output.
`weaver version` is supposed to print out the weaver module version,
something like:

```
$ weaver version
weaver v0.15.0
```

Previously, we used [`runtime.ReadBuildInfo`][ReadBuildInfo] to read the
version of the main module. However, I realized that this version was
always the string `(devel)`. At first, I thought the version was
`(devel)` when on a non-tagged commit, but later realized that it is
literally always `(devel)`: https://github.com/golang/go/issues/29228.

I did some Googling to figure out how to print out the current module
version, but it seems impossible? This PR gives up and sticks with
showing the git commit. It's not as clear, but you can look up the
commit in the repo history to find the module version.

[ReadBuildInfo]: https://pkg.go.dev/runtime/debug#ReadBuildInfo