Check for locked database when writing to perfetto db.
Previously we would see an infrequent error message saying the
database was locked. This was due to a missing isLocked check on the
path that attempts to assign a replica number to the current
weavelet. Fixed by retrying on a locking error.

Also improved a corresponding error message.

Added a benchmark for calling component methods when tracing is turned
on. (This benchmark is what would trigger the preceding issue.)