Added test to check forgotten codegen updates.
In PR #388, we changed `weaver generate` to embed listener information
in compiled Service Weaver binaries. This change landed in v0.15.0.
However, we forgot to update the codegen version. This led to the
following bug, reported by @renanbastos93 in the Discord:

- Run `weaver generate` at version v0.14.0.
- Run `go run .` at version v0.16.1.
- The app crashes looking for listener information in the binary, but
  the information is not there because of the stale `weaver generate`.

If we had updated the codegen version, then the code would not have
compiled, and the user would have received an error message explaining
that they needed to update `weaver generate`.

It's really easy to forget to update the codegen version, so I added a
unit test that checks for changes to how weaver_gen.go files are
generated. If there is any change, the test fails and you're prompted to
update the codegen version.