Implemented very rough prototype of simulator.
The Service Weaver team has been discussing more sophisticated ways of
testing Service Weaver apps. This PR implements a very rough prototype
of an idea I had. The prototype is stand-alone and does not depend on
the rest of the weaver code at all. I also don't intend for the
prototype to ever be merged in. I wrote it to further my understanding
of the idea, to prove to myself it was possible to implement something
like this, and to identify the design decisions that come up.

> Overview

The main idea of the prototype is to deterministically simulate the
execution of a distributed program with injected failures and
complex interleavings. To implement this, whenever a component calls a
method on another component, execution is halted and returned to a
central simulator. The simulator then picks the next action to take.
Specifically, the simulator can do one of the following.

    1. Spawn a brand new operation.
    2. Deliver a pending method call.
    3. Deliver a pending method return.

A more feature complete simulator could also return a failure to a
pending method call, induce a network partition, etc.

This is similar to [deterministic simulation][1] and a bit like [TLA+
model checking][2]. It is also similar to [something I implemented in
grad school][3] to test and visualize distributed algorithms. For that,
when control is returned to the central simulator, the user can pick the
next action via a web UI.

> Details

I plan on writing a design doc that details this approach to testing as
well as alternatives.

[1]: https://asatarin.github.io/testing-distributed-systems/#deterministic-simulation
[2]: https://asatarin.github.io/testing-distributed-systems/#tla
[3]: https://mwhittaker.github.io/frankenpaxos/js/src/main/js/echo/echo.html