Moved most of weaver code to internal package.
This PR moves the majority of the code in the `weaver` package into a
newly created `internal/weaver` package.

Recall that weavertests run full Service Weaver applications with a
couple of small tweaks. For example, a weavertest can provide a set of
fake component implementations that replace the real implementations.

Before this PR, there was a bit of awkwardness between the weaver
package and the weavertest package. Because the main weaver package is
public, it cannot export any functions that might be useful to the
weavertest package. To solve this, we did a couple of hacks. First, we
pass some arguments to a weavelet via a context. Second, we had a
`private` package that acted as an intermediary. To avoid cyclic
dependencies, this `private` package used `any` liberally and relied on
the callers of the package to cast to the appropriate types.

This PR is the first step in resolving the awkwardness. By moving most
of the weaver code into an internal package, we can export functions
useful to weavertest. This PR deletes the `private` package entirely.

In future PRs, I think we can continue to clean up the APIs. This PR
tries to make as few changes as possible.