Allow AutoMarshal errors to have value or pointer Error() method.
The type T passed to RegisterSerializable[T] is now always
AutoMarshal. However the error interface may be implemented by either
T or *T.

There are now two encoding tags for custom error values: one tag if
the error itself is AutoMarshal, another tag if a pointer to the error
is AutoMarshal.

errors.Is etc. do not handle cyclic errors well, so make a separate
test for cyclic errors that just verifies that we can encode and
decode cyclic errors.