Store trace spans in a relational DB format. (#459)
* Store trace spans in a relational DB format.

This format allows for traces to be queries. Here are the changes to the
trace storage format:
  * Traces are stored in encoded proto format. This format is more
    efficient than the earlier Perfetto JSON format.
  * Some trace fields are also available as table columns, for fast
    query matching.
  * Traces are converted to the Perfetto JSON format during trace
    retrieval and display.

As discussed offline with sanjay@, traces are no longer displayed in
Perfetto in bulk. Instead, each trace has a separate link and the
Perfetto UI always displays a single trace. This makes it easier to
visualize individual trace executions, as they are no longer
interleaved on the same timeline.

Other changes:
  * Fold Span proto messages into the `Span` type, as their names are
    quite generic (e.g., `Attribute`).
  * Pass trace data between weavelet and envelope in Proto format,
    since the Proto format can be efficiently serialized, if needed.
    (The OTel format doesn't support efficient serialization.)
  * Minor code re-organization.

COMPATIBILITY NOTE

This change requires a deployer version bump, as it changes the trace
wire format.

* Addressed final comments.