Slightly simplified how components are filled.
This PR replaces weaver.IsImplements, weaver.IsRef, and
weaver.IsListener with weaver.SetLogger, weaver.FillRefs, and
weaver.FillListeners.

weaver.SetLogger is able to set the logger of a weaver.Implements
without having to use reflection. For this to work, the function must be
in the root weaver package.

weaver.FillListeners is a slightly simplified version of the existing
FillListeners function. It checks if a field is a Listener by comparing
to the Listener type directly. It also sets the listener and proxy
address by casting directly to a *Listener. This function doesn't have
to be in the root weaver package, but it can't be in the internal weaver
package because of a cyclic dependency.

weaver.FillRefs is pretty much unmodified from the existing FillRefs
function, but I moved it into the root weaver package to sit next to
SetLogger and FillListeners.