Removed dependency on token package. (#474)
Removed dependency on token package.

This PR replaces token.IsIdentifier with a custom implementation to
remove the dependency on the token package.

Note that this PR now allows listener names to be go keywords, but I
think that's okay. I don't think we intended to disallow these names in
the first place.