Added bare-bones "simulator".
This PR introduces a skeleton of a determinstic simulator. The simulator
constructs a couple of replicas of every registered component. The
`weaver.Ref` fields inside the components are instantiated with reflect
stubs that, for now, just call the requested method.

The simulator in its current state isn't very useful. In the future,
I'll expand the simulator to actually simulate things. I'll also add
more comments in the future once we settle on an API. I'm also doing
everything in an internal package for now.