Register operations (ops) with simulator.
This PR introduces an `Op[T]` type to represent operations executed as
part of a simulation. Currently, the ops are not actually executed; that
will come in a future PR.

I'm not 100% satisfied with the `Op` API right now. The biggest
awkwardness comes when your operation wants to receive more than one
random argument. To do that with this API, you have to wrap all
arguments into a struct. That is clunky.

Registering an `Op` is also a bit strange because you have to call a
function rather than a `Sim` method. This is because methods in Go
cannot be generic.

I'd love to improve the API later if we can think of something better,
and I'm definitely open to suggestions.