Dim attributes in pretty printed logs.
Before this PR, log attribute names were purple and attribute values
were white. This PR dims both a light gray. This draws more attention to
the body of the log compared to the attributes.

I think we were also putting way too much stuff in the attributes which
was making the logs hard to read. In theory, you can use attributes to
query logs, but I don't think we ever did this. And even if we did, the
attributes we had were awkward to query.