Fix race in weavertest multi shutdown.
Previously, if a remote weavelet paniced, it's exit raced with
weavertest cleanup. If the code reading from the remote weavelet
detected the broken connection before weavertest got a chance to mark
the test as done, we would print an error message and exit the test
process. This interacted poorly with the weavertest/internal/generate
test that intentionally triggers a panic in a remote component.

We now just log the error and do not exit. Furthermore, the message is
written to the system logger, not to stderr, so it is suppressed
unless -test.v is supplied.

Made the race more likely to be triggered by putting a short sloop
in the test that triggers a remote panic.