Validate listener names at application startup.
This PR extends the `validateRegistrations` function to also validate
listener names when a Service Weaver application begins. As with
PR #500, this PR aims to catch permanent and easily avoidable
programming errors immediately.

I also fixed a small bug in listener name validation. Previously, if a
listener had a `weaver:""` annotation, it was being ignored. Now, it is
being rejected (since `""` is not a valid listener name).

I also updated the website documentation to describe legal listener
names.