Fixed small bug in TestRememberDraining.
I noticed that TestRememberDraining was flaky. The test closes a client
and expects all draining connections to be closed. The test had a race
though. We checked that a connection was closed, but didn't wait for the
client to finish closing. This PR changes to the test to fully close the
client and then test that all its connections are closed.