Implemented basic simulation.
This PR implements the core logic of the deterministic simulator. The
simulator maintains a buffer of pending calls and pending replies. It
repeatedly performs one of the following actions:

1. Start an op.
2. Deliver a method call.
3. Deliver a method reply.

> Starting An Op

Recall the definition of an op:

```go
type Op[T any] struct {
    Name string
    Gen func(*rand.Rand) T
    Func any
}
```

When a simulator starts an op, it calls the `Gen` function to generate a
random value. It then invokes `Func`, passing it handles to the
necessary components. These handles are generated using reflect stubs
created with the simulator's `call` method. `Func` runs freely until it
calls a component method. At that point, the reflect stub is invoked,
the arguments of the call are recorded for later execution, and the call
blocks on a channel to receive return values. When the op terminates,
the simulation is aborted if the op returns an error.

> Delivering a Method Call

When the simulator delivers a call, it picks a component replica to
execute the call and then calls the component's method directly. From
here, the executing method acts like an executing op. When the call
returns, the return values are recorded for later delivery.

> Delivering a Method Reply

When the simulator delivers a method reply, it writes the return values
of the method to the channel on which the caller is blocked. This will
signal the caller to wake up and continue executing.

> Control Flow

Most of the complexity of the simulator comes from its complicated
control flow. Execution is similar to [cooperative threading][1]. The
various ops and method calls all happen on different goroutines and are
never preempted. Instead, they relinquish control back to the simulator
whenever they call a component method or return.

[1]: https://en.wikipedia.org/wiki/Cooperative_multitasking